% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_mining.R
\name{build_business_log}
\alias{build_business_log}
\title{Business (case-level) event log from status transitions}
\usage{
build_business_log(history)
}
\arguments{
\item{history}{Status history.}
}
\value{
Data frame \code{case_id}, \code{activity}, \code{process_phase}, \code{resource},
\code{timestamp}, \code{comment}.
}
\description{
Business (case-level) event log from status transitions
}
