% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{callout}
\alias{callout}
\title{Callout box}
\usage{
callout(..., title = NULL, type = "info")
}
\arguments{
\item{...}{Content.}

\item{title}{Title.}

\item{type}{Emphasis: \code{"info"}, \code{"low"}, \code{"medium"} or \code{"high"} (darker = more urgent).}
}
\value{
HTML tag.
}
\description{
Callout box
}
