% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_roles.R
\name{can}
\alias{can}
\title{Can a role perform an action?}
\usage{
can(role, action)
}
\arguments{
\item{role}{Role (\code{"contributor"} / \code{"superuser"}), or a user list from \code{\link[=app_user]{app_user()}}.}

\item{action}{Action id (see \code{\link[=permission_matrix]{permission_matrix()}}).}
}
\value{
Logical.
}
\description{
Can a role perform an action?
}
