% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_dir}
\alias{config_dir}
\title{Directory of the CSV configuration shipped with the package}
\usage{
config_dir()
}
\value{
A path.
}
\description{
Directory of the CSV configuration shipped with the package
}
