% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_users.R
\name{connect_users}
\alias{connect_users}
\title{Users from the Posit Connect Server API (all pages)}
\usage{
connect_users(server, key, page_size = 500)
}
\arguments{
\item{server}{Connect URL.}

\item{key}{API key.}

\item{page_size}{Page size.}
}
\value{
Data frame \code{username}, \code{name}, \code{email}.
}
\description{
Users from the Posit Connect Server API (all pages)
}
