% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_activate_value_model}
\alias{db_activate_value_model}
\title{Re-activate a previously published model version (rollback)}
\usage{
db_activate_value_model(con, model_id)
}
\arguments{
\item{con}{A DBI connection.}

\item{model_id}{Model id.}
}
\description{
Re-activate a previously published model version (rollback)
}
