% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_delete_m4_line}
\alias{db_delete_m4_line}
\title{Delete a 4MC calculation line}
\usage{
db_delete_m4_line(con, line_id)
}
\arguments{
\item{con}{A DBI connection.}

\item{line_id}{Line id.}
}
\description{
Delete a 4MC calculation line
}
