% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect (and shut down an embedded DuckDB instance)}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\description{
Disconnect (and shut down an embedded DuckDB instance)
}
