% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_audits}
\alias{db_get_audits}
\title{Read value audits}
\usage{
db_get_audits(con, id = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Optional initiative id.}
}
\value{
Data frame.
}
\description{
Read value audits
}
