% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_initiatives}
\alias{db_get_initiatives}
\title{Read initiatives}
\usage{
db_get_initiatives(con, id = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Optional id filter.}
}
\value{
Data frame.
}
\description{
Read initiatives
}
