% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_m4_lines}
\alias{db_get_m4_lines}
\title{Read 4MC calculation lines}
\usage{
db_get_m4_lines(con, id = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Optional initiative id.}
}
\value{
Data frame.
}
\description{
Read 4MC calculation lines
}
