% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_prmt_lines}
\alias{db_get_prmt_lines}
\title{Read PRMT calculation lines}
\usage{
db_get_prmt_lines(con, id = NULL, stage = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Optional initiative id.}

\item{stage}{Optional stage filter.}
}
\value{
Data frame.
}
\description{
Read PRMT calculation lines
}
