% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_reviews}
\alias{db_get_reviews}
\title{Read expert reviews}
\usage{
db_get_reviews(con, id = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Optional initiative id.}
}
\value{
Data frame.
}
\description{
Read expert reviews
}
