% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_get_status_history}
\alias{db_get_status_history}
\title{Status history}
\usage{
db_get_status_history(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
Data frame.
}
\description{
Status history
}
