% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_init}
\alias{db_init}
\title{Create the schema if it does not exist}
\usage{
db_init(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
Invisibly, the table names.
}
\description{
Create the schema if it does not exist
}
