% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_next_initiative_id}
\alias{db_next_initiative_id}
\title{Next human-readable initiative id (DV-0001, DV-0002, ...)}
\usage{
db_next_initiative_id(con)
}
\arguments{
\item{con}{A DBI connection.}
}
\value{
Character id.
}
\description{
Next human-readable initiative id (DV-0001, DV-0002, ...)
}
