% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_publish_value_model}
\alias{db_publish_value_model}
\title{Publish a calibrated value model (new active version)}
\usage{
db_publish_value_model(
  con,
  record,
  comment = NA,
  user = "unknown",
  time = Sys.time()
)
}
\arguments{
\item{con}{A DBI connection.}

\item{record}{Output of \code{\link[=model_record]{model_record()}}.}

\item{comment}{Calibration comment.}

\item{user}{Superuser publishing the model.}

\item{time}{Publication time.}
}
\value{
The model id.
}
\description{
Previous versions of the same kind are kept but deactivated.
}
