% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_request_validation}
\alias{db_request_validation}
\title{Send the 4MC valuation of an initiative to a validator}
\usage{
db_request_validation(
  con,
  id,
  validator,
  user = "unknown",
  comment = NA,
  time = Sys.time()
)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Initiative id.}

\item{validator}{User name of the validator (Posit Connect user).}

\item{user}{User sending the request.}

\item{comment}{Optional note to the validator.}

\item{time}{Request time.}
}
\value{
The validation id.
}
\description{
Any pending request of the initiative is superseded by the new one.
}
