% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_save_rice}
\alias{db_save_rice}
\title{Save (upsert) the RICE assessment of an initiative}
\usage{
db_save_rice(con, id, rice, user = "unknown", time = Sys.time())
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Initiative id.}

\item{rice}{List with \code{users}, \code{impact}, \code{confidence}, \code{effort},
\code{reach_value}, \code{score}, \code{rationale}.}

\item{user}{User.}

\item{time}{Scoring time.}
}
\description{
Save (upsert) the RICE assessment of an initiative
}
