% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_set_status}
\alias{db_set_status}
\title{Change the status of an initiative and record it in the history}
\usage{
db_set_status(
  con,
  id,
  status,
  user = "unknown",
  comment = NA,
  time = Sys.time()
)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Initiative id.}

\item{status}{New status.}

\item{user}{User.}

\item{comment}{Optional comment.}

\item{time}{Change time.}
}
\value{
\code{TRUE} if the status changed.
}
\description{
Change the status of an initiative and record it in the history
}
