% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{db_update_initiative}
\alias{db_update_initiative}
\title{Update editable fields of an initiative}
\usage{
db_update_initiative(con, id, fields)
}
\arguments{
\item{con}{A DBI connection.}

\item{id}{Initiative id.}

\item{fields}{Named list of columns to update.}
}
\value{
Number of rows updated.
}
\description{
Update editable fields of an initiative
}
