% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{echart_from_option}
\alias{echart_from_option}
\title{Render an ECharts option with echarts4r}
\usage{
echart_from_option(option, height = NULL)
}
\arguments{
\item{option}{An ECharts option list.}

\item{height}{CSS height.}
}
\value{
An echarts4r htmlwidget.
}
\description{
Render an ECharts option with echarts4r
}
