% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{fmt_num}
\alias{fmt_num}
\title{Format a number for display}
\usage{
fmt_num(x, digits = 1, suffix = "")
}
\arguments{
\item{x}{Number.}

\item{digits}{Decimals.}

\item{suffix}{Unit suffix.}
}
\value{
Character.
}
\description{
Format a number for display
}
