% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{group_colors}
\alias{group_colors}
\title{Group colours of the prioritisation scatter}
\usage{
group_colors()
}
\value{
Named character vector.
}
\description{
Group colours of the prioritisation scatter
}
