% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_roles.R
\name{is_owner}
\alias{is_owner}
\title{Is the user the owner (or creator) of an initiative?}
\usage{
is_owner(usr, row)
}
\arguments{
\item{usr}{User list from \code{\link[=app_user]{app_user()}}.}

\item{row}{Initiative row (with \code{owner}, \code{created_by}).}
}
\value{
Logical.
}
\description{
Is the user the owner (or creator) of an initiative?
}
