% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_ui.R
\name{m4_inputs}
\alias{m4_inputs}
\title{Inputs for a set of 4MC figures (shared by expert review and audit)}
\usage{
m4_inputs(ns)
}
\arguments{
\item{ns}{Namespace function.}
}
\value{
UI tags.
}
\description{
Inputs for a set of 4MC figures (shared by expert review and audit)
}
