% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_4m.R
\docType{data}
\name{m4_value_metrics}
\alias{m4_value_metrics}
\title{Codes of the value metrics (excluding cost)}
\format{
An object of class \code{character} of length 4.
}
\usage{
m4_value_metrics
}
\description{
Codes of the value metrics (excluding cost)
}
\keyword{datasets}
