% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_methodology.R
\docType{data}
\name{methodology_languages}
\alias{methodology_languages}
\title{Languages of the methodology page}
\format{
An object of class \code{character} of length 2.
}
\usage{
methodology_languages
}
\description{
Languages of the methodology page
}
\keyword{datasets}
