% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_audit_server.R
\name{mod_audit_server}
\alias{mod_audit_server}
\title{Execution and audit server}
\usage{
mod_audit_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Execution and audit server
}
