% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_initiative_server.R
\name{mod_initiative_server}
\alias{mod_initiative_server}
\title{Initiative tab server}
\usage{
mod_initiative_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state (see \code{\link[=app_server]{app_server()}}).}
}
\description{
Initiative tab server
}
