% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology_server.R
\name{mod_methodology_server}
\alias{mod_methodology_server}
\title{Methodology tab server}
\usage{
mod_methodology_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Methodology tab server
}
