% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology_ui.R
\name{mod_methodology_ui}
\alias{mod_methodology_ui}
\title{Methodology tab UI (English / Spanish)}
\usage{
mod_methodology_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Methodology tab UI (English / Spanish)
}
