% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_models_server.R
\name{mod_models_server}
\alias{mod_models_server}
\title{Value models (admin) server}
\usage{
mod_models_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Value models (admin) server
}
