% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_models_ui.R
\name{mod_models_ui}
\alias{mod_models_ui}
\title{Value models (admin) UI: calibrate, compare and publish}
\usage{
mod_models_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Value models (admin) UI: calibrate, compare and publish
}
