% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview_server.R
\name{mod_overview_server}
\alias{mod_overview_server}
\title{Portfolio overview server}
\usage{
mod_overview_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Portfolio overview server
}
