% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parameters_server.R
\name{mod_parameters_server}
\alias{mod_parameters_server}
\title{Parameters server}
\usage{
mod_parameters_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Parameters server
}
