% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phase1_server.R
\name{mod_phase1_server}
\alias{mod_phase1_server}
\title{Phase I (RICE) server}
\usage{
mod_phase1_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Phase I (RICE) server
}
