% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phase2_server.R
\name{mod_phase2_server}
\alias{mod_phase2_server}
\title{Phase II (PRMT monetary valuation) server}
\usage{
mod_phase2_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Phase II (PRMT monetary valuation) server
}
