% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phase2_ui.R
\name{mod_phase2_ui}
\alias{mod_phase2_ui}
\title{Phase II (PRMT monetary valuation) UI}
\usage{
mod_phase2_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Phase II (PRMT monetary valuation) UI
}
