% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phase3_server.R
\name{mod_phase3_server}
\alias{mod_phase3_server}
\title{Phase III (planning review) server}
\usage{
mod_phase3_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Phase III (planning review) server
}
