% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_portfolio_server.R
\name{mod_portfolio_server}
\alias{mod_portfolio_server}
\title{Portfolio tab server}
\usage{
mod_portfolio_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Portfolio tab server
}
