% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_process_server.R
\name{mod_process_server}
\alias{mod_process_server}
\title{Process log (process mining) server}
\usage{
mod_process_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Process log (process mining) server
}
