% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_register_ui.R
\name{mod_register_ui}
\alias{mod_register_ui}
\title{Registration & RICE UI (step 1, recorded at once)}
\usage{
mod_register_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Registration & RICE UI (step 1, recorded at once)
}
