% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_server.R
\name{mod_review_server}
\alias{mod_review_server}
\title{Expert review server}
\usage{
mod_review_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Expert review server
}
