% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_ui.R
\name{mod_review_ui}
\alias{mod_review_ui}
\title{Expert review UI (step 3, manual evaluation with the same 4MC metrics)}
\usage{
mod_review_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Expert review UI (step 3, manual evaluation with the same 4MC metrics)
}
