% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_server.R
\name{mod_sidebar_server}
\alias{mod_sidebar_server}
\title{Sidebar server}
\usage{
mod_sidebar_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state (see \code{\link[=app_server]{app_server()}}).}
}
\description{
Sidebar server
}
