% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar_ui.R
\name{mod_sidebar_ui}
\alias{mod_sidebar_ui}
\title{Sidebar UI: initiative selector, process stepper and workflow actions}
\usage{
mod_sidebar_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Sidebar UI: initiative selector, process stepper and workflow actions
}
