% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_model_server.R
\name{mod_value_model_server}
\alias{mod_value_model_server}
\title{Value model (RICE to PRMT regression) server}
\usage{
mod_value_model_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Value model (RICE to PRMT regression) server
}
