% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_model_ui.R
\name{mod_value_model_ui}
\alias{mod_value_model_ui}
\title{Value model (RICE to PRMT regression) UI}
\usage{
mod_value_model_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Value model (RICE to PRMT regression) UI
}
