% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\docType{data}
\name{mono}
\alias{mono}
\title{Monochrome palette (slate), from lightest (\code{c050}) to darkest (\code{c900})}
\format{
An object of class \code{list} of length 10.
}
\usage{
mono
}
\description{
Monochrome palette (slate), from lightest (\code{c050}) to darkest (\code{c900})
}
\keyword{datasets}
