% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_db.R
\name{now_utc}
\alias{now_utc}
\title{Current UTC timestamp as ISO text}
\usage{
now_utc(time = Sys.time())
}
\arguments{
\item{time}{A POSIXct.}
}
\value{
Character.
}
\description{
Current UTC timestamp as ISO text
}
