% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_roles.R
\name{permission_matrix}
\alias{permission_matrix}
\title{Permission matrix}
\usage{
permission_matrix()
}
\value{
Data frame \code{action}, \code{label}, \code{contributor}, \code{superuser}.
}
\description{
Permission matrix
}
