% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plan_actual_option}
\alias{plan_actual_option}
\title{ECharts option: planned vs actual value per PRMT metric}
\usage{
plan_actual_option(pva)
}
\arguments{
\item{pva}{Output of \code{\link[=plan_vs_actual]{plan_vs_actual()}}.}
}
\value{
An ECharts option list.
}
\description{
ECharts option: planned vs actual value per PRMT metric
}
