% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_group}
\alias{plot_group}
\title{Chart group of a status (prioritisation scatter)}
\usage{
plot_group(status)
}
\arguments{
\item{status}{Status vector.}
}
\value{
Character vector of groups.
}
\description{
Chart group of a status (prioritisation scatter)
}
