% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rice.R
\name{reach_value}
\alias{reach_value}
\title{Reach value: log10 of the number of users reached}
\usage{
reach_value(users)
}
\arguments{
\item{users}{Number of users.}
}
\value{
Numeric vector.
}
\description{
Values below 1 user are floored to 1 (reach value 0).
}
