% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rice.R
\name{rice_ordinal}
\alias{rice_ordinal}
\title{Ordinal position of RICE levels (1 = smallest)}
\usage{
rice_ordinal(level, dimension, cfg)
}
\arguments{
\item{level}{Character vector of levels.}

\item{dimension}{Dimension name.}

\item{cfg}{Configuration list.}
}
\value{
Integer vector.
}
\description{
Ordinal position of RICE levels (1 = smallest)
}
