% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rice.R
\name{rice_weight}
\alias{rice_weight}
\title{Numeric weight of RICE levels}
\usage{
rice_weight(level, dimension, cfg)
}
\arguments{
\item{level}{Character vector of levels.}

\item{dimension}{Dimension name.}

\item{cfg}{Configuration list.}
}
\value{
Numeric vector (\code{NA} for unknown levels).
}
\description{
Numeric weight of RICE levels
}
