% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gates.R
\docType{data}
\name{status_assessment}
\alias{status_assessment}
\alias{status_decision}
\alias{status_all}
\title{Initiative statuses}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 9.
}
\usage{
status_assessment

status_decision

status_all
}
\description{
Assessment statuses are derived from the data (what is still pending);
decision statuses are set explicitly by a workflow action.
}
\keyword{datasets}
