% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{status_badge}
\alias{status_badge}
\title{Status badge coloured as in the charts}
\usage{
status_badge(status)
}
\arguments{
\item{status}{Status.}
}
\value{
HTML tag.
}
\description{
Status badge coloured as in the charts
}
