% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_4m.R
\name{time_value_usd_hour}
\alias{time_value_usd_hour}
\title{Hourly value of saved time}
\usage{
time_value_usd_hour(cfg)
}
\arguments{
\item{cfg}{Configuration list.}
}
\value{
USD per hour.
}
\description{
Average salary per hour x productivity coefficient (value generated when
the saved time is redeployed to productive tasks).
}
