% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_ui.R
\name{update_m4_inputs}
\alias{update_m4_inputs}
\title{Fill 4MC inputs from a list of figures}
\usage{
update_m4_inputs(session, v)
}
\arguments{
\item{session}{Module session.}

\item{v}{Named list \code{P}, \code{R}, \code{M}, \code{T}, \code{C}, \code{category}.}
}
\description{
Fill 4MC inputs from a list of figures
}
