% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_users.R
\name{user_choices}
\alias{user_choices}
\title{Choices for a user selector ("Name (username)" -> username)}
\usage{
user_choices(users, extra = character(0))
}
\arguments{
\item{users}{Output of \code{\link[=user_directory]{user_directory()}}.}

\item{extra}{Usernames to include even if not in the directory.}
}
\value{
Named character vector.
}
\description{
Choices for a user selector ("Name (username)" -> username)
}
