% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_users.R
\name{user_directory}
\alias{user_directory}
\title{Read the user directory}
\usage{
user_directory(refresh = FALSE)
}
\arguments{
\item{refresh}{Ignore the cache.}
}
\value{
Data frame \code{username}, \code{name}, \code{email}.
}
\description{
Read the user directory
}
