% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_users.R
\name{user_label}
\alias{user_label}
\title{Display name of users}
\usage{
user_label(username, users)
}
\arguments{
\item{username}{User names.}

\item{users}{Output of \code{\link[=user_directory]{user_directory()}}.}
}
\value{
Character vector (falls back to the user name).
}
\description{
Display name of users
}
