% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_roles.R
\name{user_role}
\alias{user_role}
\title{Resolve the role of a user}
\usage{
user_role(user, groups = character(0))
}
\arguments{
\item{user}{User name.}

\item{groups}{Groups of the user (Posit Connect).}
}
\value{
\code{"superuser"} or \code{"contributor"}.
}
\description{
Resolve the role of a user
}
