% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_value_model.R
\name{value_fit_option}
\alias{value_fit_option}
\title{ECharts option: predicted vs actual (multivariable model)}
\usage{
value_fit_option(model)
}
\arguments{
\item{model}{Output of \code{\link[=fit_value_model]{fit_value_model()}}.}
}
\value{
An ECharts option list.
}
\description{
ECharts option: predicted vs actual (multivariable model)
}
