% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_value_model.R
\name{value_model_features}
\alias{value_model_features}
\title{Model features from RICE inputs}
\usage{
value_model_features(users, impact, confidence, effort, cfg)
}
\arguments{
\item{users, impact, confidence, effort}{RICE inputs.}

\item{cfg}{Configuration list.}
}
\value{
Data frame of features.
}
\description{
Model features from RICE inputs
}
