% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_value_model.R
\docType{data}
\name{value_model_kinds}
\alias{value_model_kinds}
\title{Value model kinds}
\format{
An object of class \code{character} of length 2.
}
\usage{
value_model_kinds
}
\description{
Value model kinds
}
\keyword{datasets}
