% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{app_user}
\alias{app_user}
\title{Identify the user (Posit Connect aware)}
\usage{
app_user(session)
}
\arguments{
\item{session}{Shiny session.}
}
\value{
List \code{user}, \code{groups}, \code{is_planning}.
}
\description{
On Posit Connect \code{session$user} and \code{session$groups} are populated from the
authenticated identity; locally the OS user is used.
}
