% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{callout}
\alias{callout}
\title{Callout box}
\usage{
callout(..., title = NULL, type = "info")
}
\arguments{
\item{...}{Content.}

\item{title}{Title.}

\item{type}{One of \code{"info"}, \code{"success"}, \code{"warning"}, \code{"danger"}.}
}
\value{
HTML tag.
}
\description{
Callout box
}
