% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_audit_ui.R
\name{mod_audit_ui}
\alias{mod_audit_ui}
\title{Value audit UI (realisation, post-execution): actual 4M plus adoption}
\usage{
mod_audit_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Value audit UI (realisation, post-execution): actual 4M plus adoption
}
