% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_register_server.R
\name{mod_register_server}
\alias{mod_register_server}
\title{Registration server}
\usage{
mod_register_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
Registration server
}
