% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_valuation_server.R
\name{mod_valuation_server}
\alias{mod_valuation_server}
\title{4MC valuation server}
\usage{
mod_valuation_server(id, state)
}
\arguments{
\item{id}{Module id.}

\item{state}{Shared application state.}
}
\description{
4MC valuation server
}
