% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gates.R
\docType{data}
\name{review_decisions}
\alias{review_decisions}
\title{Planning review decisions}
\format{
An object of class \code{character} of length 3.
}
\usage{
review_decisions
}
\description{
Planning review decisions
}
\keyword{datasets}
