% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gates.R
\docType{data}
\name{status_assessment}
\alias{status_assessment}
\alias{status_decision}
\alias{status_all}
\title{Initiative statuses}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.
}
\usage{
status_assessment

status_decision

status_all
}
\description{
\code{Recorded} and \code{Evaluated} are derived from the data (are the required
evaluations complete?); \code{Delivered} and \code{Audited} are set by an action.
}
\keyword{datasets}
